Fake_BSOD/requirements.txt
pillow>=9.0.0
qrcode[pil]>=7.3